Register an alembic dialect
Alembic is the migration add-on for SQLAlchemy with its own dialect
system. Detect the presence of this package and register a dialect
(currently just an alias for the postgres dialect).

Alembic's test suite is not cleanly factored for us to reuse, but in
manual testing it appears to work except for two issues:
- For tables without a primary key, we create a hidden `rowid` column.
  But it's apparently not hidden enough; alembic sees it and gets
  confused. This appears to affect a lot of alembic's tests so I'm not
  sure how much coverage we're actually getting here.
- We don't yet support ON DELETE or ON UPDATE clauses for foreign keys.