Hide the primary index from SQLAlchemy and its consumers
This caused a problem with applications that used reflection to compare
the models against the database state, for example Alembic. Alembic
attempts to autogenerate migrations that forward the database to match
models. What this means is that the implicit relations such as the
primary index will then be scheduled to drop.