Update all the versions
We were only testing CRDB versions 1.0 and a beta of 1.1, and using
python 3.6, while the current release is 3.7. I decided to upgrade
python instead of installing python 3.6, which led to a yak shave:
- SQLAlchemy 1.1's test suite fails on python 3.7 due to deprecation
warnings, so upgrade that too.
- SQLAlchemy 1.2's test suite only works with CRDB 2.1-alpha due to
the LIKE ESCAPE syntax, and needs some other tweaks to the test
configs.
- CircleCI is dropping support for their V1 config format soon, so
might as well update that while we're here.

The circle scripts now work for local testing on macos.